Update AppShell, enhance app icon, and adjust project file
This commit modifies the `AppShell.xaml` file to replace the existing "Settings" page with a new "Detailed Analysis" page and adds a separate invisible `FlyoutItem` for "Settings."

In `appicon.svg`, numerous new path elements with various fill colors and opacities have been added, enhancing the visual design of the app icon.

Additionally, the `ScoreCard.csproj` file has been updated to include a `<CopyToOutputDirectory>` element for the app icon, ensuring the latest version is copied to the output directory during the build process.